requirements.txt
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
gunicorn==21.2.0
whitenoise==6.5.0
django-cors-headers==4.3.0
django-environ==0.10.0
